requirements.txt
beautifulsoup4
pandas
python-dotenv
selenium